pyproject.toml
[project]
name = "taxi-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "db-dtypes>=1.4.1",
    "google-cloud-bigquery>=3.29.0",
    "matplotlib>=3.10.0",
    "polars>=1.22.0",
    "seaborn>=0.13.2",
    "streamlit>=1.42.0",
]
